+++ Assets/Personal/Gustav/EmissionMaterial.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c5c6680e634b996418ccc9af07c295ac
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
