+++ Assets/Personal/Gustav/LightSettings.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 229fdbe09cdc0344a816e0cac4ee15ce
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
