+++ Assets/Personal/Gustav/LightSettings.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e9c7eb2a88fb94940a348229dd8326e7
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
