+++ Assets/Personal/Gustav/SunController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1271f630a7e4f154e94df947e4826d17
+guid: 9101f54204d269d45b0a9cbe93f4e311
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
