+++ Assets/Scripts/SunController.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 47a78225f9dcd5d4e961f5ebb7c2016c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
