+++ Assets/Scripts/SunController/SunController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5b577b8afc3718c49b5fb3af4d7ced31
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
