+++ Assets/Scripts/SunController/SunSetting.cs
@@ -1,0 +1,5 @@
+using UnityEngine;
+public class SunSetting : ScriptableObject {
+   public float DegreesX;
+   public float DegreesY;
+}
