+++ Assets/Scripts/SunController/SunSettings.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d01218fd3cb8ce54abe28679cbd878dc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
