+++ Assets/Scripts/SunController/SunVolumeProfile.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2ce8cfbe75ced8c4eab657d6556adf94
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
