+++ Assets/Sky and Fog Settings Profile.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e28a7a0070b5d3a4aad24a1563a66239
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
